Refactored the database code moved routes to routes.rs
