+++ Cargo.toml
@@ -10,6 +10,7 @@ actix-cors = "0.6.4"
 actix-web = "4.4.0"
 argon2 = "0.5.2"
 chrono = { version = "0.4.31", features = ["serde"] }
+deadpool-diesel = { version = "0.5.0", features = ["postgres"] }
 diesel = { version = "2.1.3", features = ["postgres", "r2d2"] }
 dotenv = "0.15.0"
 env_logger = "0.10.0"
